requirements.txt
# Core web framework
fastapi>=0.115.0
uvicorn>=0.22.0
pydantic>=2.10.0

# Advanced retrieval components
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.8.0
numpy>=1.24.0
scipy>=1.11.0

# Cross-encoder reranking (included in sentence-transformers)
# cross-encoder>=0.1.0

# Advanced document processing
PyMuPDF>=1.23.0
pdfminer.six>=20231228
pdfplumber>=0.9.0
python-docx>=1.1.0
pytesseract>=0.3.10  # OCR for scanned PDFs
Pillow>=10.0.0
ebooklib>=0.18      # EPUB processing
striprtf>=0.0.26    # RTF processing
beautifulsoup4>=4.12.0  # HTML processing

# Text processing and NLP
nltk>=3.8.1
spacy>=3.7.0
regex>=2023.10.3
chardet>=5.2.0
langdetect>=1.0.9

# Storage and indexing
h5py>=3.10.0
joblib>=1.3.2
lmdb>=1.4.1
pickle5>=0.0.12

# Evaluation and metrics
scikit-learn>=1.3.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
click>=8.1.7
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Optional: OpenAI integration
openai>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0